bot/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
python-telegram-bot==21.4
itsdangerous==2.2.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
